feature: implement pagination component and update comment/user management scripts
